+++ lib/generators/federails/copy_client_policies/copy_client_policies_generator.rb
@@ -1,0 +1,9 @@
+module Federails
+  class CopyClientPoliciesGenerator < Rails::Generators::Base
+    source_root File.expand_path('../../../../app/policies/federails', __dir__)
+
+    def copy_policies
+      directory 'client', Rails.root.join('app', 'policies', 'federails', 'client')
+    end
+  end
+end